moved extract logic to a separate lib
