+++ src/commands/check.ts
@@ -1,0 +1,7 @@
+async function check(pofile: string, paths: string[], locale: string) {
+    console.log(`Checking translations from ${paths}`);
+    console.log(`pofile ${pofile}`);
+    console.log(`locale ${locale}`);
+}
+
+export default check;
